Moved view to separate package
